PsychometricsBazzarBase.jl

This module provides a base for the libraries in the JuliaPsychometricsBazzar org. It contains abstractions over basic mathematical techniques such as numerical integration, optimization and interpolation. It also contains mathematical traits e.g. for function's domain.

Ideally, the package will be transitional, since functionality may make its way into more specific packages (including existing packages) over time.

PsychometricsBazzarBase.Domains

PsychometricsBazzarBase.IntegralCoeffs

PsychometricsBazzarBase.Domains — Module

This module provides some basic traits related to function domains

PsychometricsBazzarBase.IntegralCoeffs — Module

These are helpers for weighting integrals of p.d.f.s for calculating basic stats.

The main idea of doing it this way is to have a single instance of these to reuse specializations and to use structs so as to be able to control the level of specialization.