ExtraDistributions

PsychometricsBazzarBase.ExtraDistributions — Module

This module contains extra distributions for psychometrics. Currently this is limited to NormalScaledLogistic.

PsychometricsBazzarBase.ExtraDistributions.NormalScaledLogistic — Type

The normal scaled logistic distribution is an approximation to the normal distribution based upon the logistic distribution. It has been commonly used in IRT modelling, such as in the mirt package for R.

Index

PsychometricsBazzarBase.ExtraDistributions

PsychometricsBazzarBase.ExtraDistributions.NormalScaledLogistic