IntegralCoeffs

PsychometricsBazzarBase.IntegralCoeffs — Module

These are helpers for weighting integrals of p.d.f.s for calculating basic stats.

The main idea of doing it this way is to have a single instance of these to reuse specializations and to use structs so as to be able to control the level of specialization.

PsychometricsBazzarBase.IntegralCoeffs.PriorApply — Type

PsychometricsBazzarBase.IntegralCoeffs.SqDev — Type

PsychometricsBazzarBase.IntegralCoeffs.id — Method

id(x)

PsychometricsBazzarBase.IntegralCoeffs.one — Method

one(x_)

Index

PsychometricsBazzarBase.IntegralCoeffs

PsychometricsBazzarBase.IntegralCoeffs.PriorApply

PsychometricsBazzarBase.IntegralCoeffs.SqDev

PsychometricsBazzarBase.IntegralCoeffs.id

PsychometricsBazzarBase.IntegralCoeffs.one