Integrators

PsychometricsBazaarBase.Integrators — Module

This module provides a common interface to different numerical integration techniques.

PsychometricsBazaarBase.Integrators.BareIntegrationResult — Type

The result of an integration technique which provides no error value.

vec::Any

PsychometricsBazaarBase.Integrators.CubaCuhre — Type

The Cuhre algorithm.

struct CubaCuhre <: PsychometricsBazaarBase.Integrators.CubaAlgorithm

PsychometricsBazaarBase.Integrators.CubaDivonne — Type

The Divonne algorithm.

struct CubaDivonne <: PsychometricsBazaarBase.Integrators.CubaAlgorithm

PsychometricsBazaarBase.Integrators.CubaIntegrator — Type

CubaIntegrator is a wrapper around the Cuba.jl integration functions.

struct CubaIntegrator{AlgorithmT<:PsychometricsBazaarBase.Integrators.CubaAlgorithm, KwargsT} <: PsychometricsBazaarBase.Integrators.Integrator

lo::Vector{Float64}

hi::Vector{Float64}

algorithm::PsychometricsBazaarBase.Integrators.CubaAlgorithm

kwargs::Any

Usage example:

CubaIntegrator([0.0, 0.0], [1.0, 1.0], CubaVegas()) do x
    x[1] * x[2]
end

PsychometricsBazaarBase.Integrators.CubaIntegrator — Method

(integrator::CubaIntegrator)(f[, ncomp, lo, hi; kwargs...])

Perform a Cuba integration.

PsychometricsBazaarBase.Integrators.CubaSuave — Type

The Sauve algorithm.

struct CubaSuave <: PsychometricsBazaarBase.Integrators.CubaAlgorithm

PsychometricsBazaarBase.Integrators.CubaVegas — Type

The VEGAS algorithm.

struct CubaVegas <: PsychometricsBazaarBase.Integrators.CubaAlgorithm

PsychometricsBazaarBase.Integrators.CubatureIntegrator — Type

Construct a Cubature integrator based on Cubature.jl with on a specified interval.

lo::Vector{Float64}

hi::Vector{Float64}

kwargs::Any

PsychometricsBazaarBase.Integrators.CubatureIntegrator — Method

(integrator::CubatureIntegrator)(f[, ncomp, lo, hi; kwargs...])

Perform a Cubature integration based on Cubature.jl.

PsychometricsBazaarBase.Integrators.ErrorIntegrationResult — Type

The result of an integration technique which provides an error value. Note that error values are not comparible between different integration techniques in general.

vec::Any

err::Any

PsychometricsBazaarBase.Integrators.FixedGKIntegrator — Type

Construct a fixed-order Gauss-Kronrod integrator based on QuadGK.jl with on a specified interval.

lo::Float64

hi::Float64

order::Int64

PsychometricsBazaarBase.Integrators.FixedGKIntegrator — Method

(integrator::QuadGKIntegrator)(f[, ncomp, lo, hi; order=...])

Perform a fixed-order Gauss-Kronrod integration based on QuadGK.jl.

PsychometricsBazaarBase.Integrators.HCubatureIntegrator — Type

Construct a Cubature integrator based on HCubature.jl with on a specified interval.

lo::Vector{Float64}

hi::Vector{Float64}

kwargs::Any

PsychometricsBazaarBase.Integrators.HCubatureIntegrator — Method

(integrator::HCubatureIntegrator)(f[, ncomp, lo, hi; kwargs...])

Perform Cubature integration based on HCubature.jl.

PsychometricsBazaarBase.Integrators.MCIVegasIntegrator — Type

Construct a VEGAS integrator based on MonteCarloIntegration.jl with on a specified interval.

lo::Vector{Float64}

hi::Vector{Float64}

kwargs::NamedTuple

PsychometricsBazaarBase.Integrators.MCIVegasIntegrator — Method

(integrator::HCubatureIntegrator)(f[, ncomp, lo, hi; kwargs...])

Perform a VEGAS integration based on MonteCarloIntegration.jl.

PsychometricsBazaarBase.Integrators.MultiDimFixedGKIntegrator — Type

Construct a fixed-order multi-dimensional Gauss-Kronrod integrator based on QuadGK.jl with on a specified interval.

lo::Vector{Float64}

hi::Vector{Float64}

order::AbstractVector{Int64}

PsychometricsBazaarBase.Integrators.MultiDimFixedGKIntegrator — Method

(integrator::QuadGKIntegrator)(f[, ncomp, lo, hi; order=...])

Perform a fixed-order multi-dimensional Gauss-Kronrod integrator based on QuadGK.jl.

PsychometricsBazaarBase.Integrators.QuadGKIntegrator — Type

Construct a adaptive Gauss-Kronrod integrator based on QuadGK.jl with on a specified interval.

lo::Float64

hi::Float64

order::Int64

PsychometricsBazaarBase.Integrators.QuadGKIntegrator — Method

(integrator::QuadGKIntegrator)(f[, ncomp, lo, hi; order=..., rtol=...])

Perform an adaptive Gauss-Kronrod integration using QuadGK.jl.

PsychometricsBazaarBase.Integrators.interr — Function

Given any integration result, get the integral error. In case the integration technique does not supply one, this returns nothing.

PsychometricsBazaarBase.Integrators.intval — Method

Given any integration result, get the integral value.

intval(res)

Index

PsychometricsBazaarBase.Integrators

PsychometricsBazaarBase.Integrators.BareIntegrationResult

PsychometricsBazaarBase.Integrators.CubaCuhre

PsychometricsBazaarBase.Integrators.CubaDivonne

PsychometricsBazaarBase.Integrators.CubaIntegrator

PsychometricsBazaarBase.Integrators.CubaIntegrator

PsychometricsBazaarBase.Integrators.CubaSuave

PsychometricsBazaarBase.Integrators.CubaVegas

PsychometricsBazaarBase.Integrators.CubatureIntegrator

PsychometricsBazaarBase.Integrators.CubatureIntegrator

PsychometricsBazaarBase.Integrators.ErrorIntegrationResult

PsychometricsBazaarBase.Integrators.FixedGKIntegrator

PsychometricsBazaarBase.Integrators.FixedGKIntegrator

PsychometricsBazaarBase.Integrators.HCubatureIntegrator

PsychometricsBazaarBase.Integrators.HCubatureIntegrator

PsychometricsBazaarBase.Integrators.MCIVegasIntegrator

PsychometricsBazaarBase.Integrators.MCIVegasIntegrator

PsychometricsBazaarBase.Integrators.MultiDimFixedGKIntegrator

PsychometricsBazaarBase.Integrators.MultiDimFixedGKIntegrator

PsychometricsBazaarBase.Integrators.QuadGKIntegrator

PsychometricsBazaarBase.Integrators.QuadGKIntegrator

PsychometricsBazaarBase.Integrators.interr

PsychometricsBazaarBase.Integrators.intval