ConfigTools

PsychometricsBazaarBase.ConfigTools — Module

This module contains utilities to implement highly configurible library code where configuration is performed through structs, and smart defaults allow sloppy or flat specification of otherwise deeply nested configuration structs.

PsychometricsBazaarBase.ConfigTools.find1 — Method

find1(pred, iter, fail_msg)

Given an iterable iter and a predicate pred, this function returns a match or else nothing if no match. In case there are multiple matches, an error is thrown.

PsychometricsBazaarBase.ConfigTools.find1_instance — Method

find1_instance(type, iter)

Returns exactly one instance in iter of type type or else nothing`. In case there are multiple matches, an error is thrown.

PsychometricsBazaarBase.ConfigTools.find1_type — Method

find1_type(type, iter)

Returns exactly one type in iter of which is a subtype of type or else nothing`. In case there are multiple matches, an error is thrown.

PsychometricsBazaarBase.ConfigTools.find1_type_sloppy — Method

find1_type_sloppy(type, iter)

Returns exactly one type in iter of which is either a subtype of type or an instance of type or else nothing`. In case there are multiple matches, an error is thrown.

PsychometricsBazaarBase.ConfigTools.@requiresome — Macro

This macro is passed an assignment like so

@requiresome foo = bar()

If bar() returns nothing, then the macro causes the current function to return nothing. Otherwise, execution continues.

PsychometricsBazaarBase.ConfigTools.@returnsome — Macro

This macro is passed an expression and a function like so

@returnsome foo() do x
    bar(x)
end

If foo() return any value apart from nothing, the macro executes the function and returns the value as long as it is not nothing. In all other cases, execution continues.

PsychometricsBazaarBase.ConfigTools.@returnsome — Macro

This macro is passed an expression like so

@returnsome foo()

If foo() return any value apart from nothing, the macro causes the current function to return that value. Otherwise, execution continues.

Index

PsychometricsBazaarBase.ConfigTools

PsychometricsBazaarBase.ConfigTools.find1

PsychometricsBazaarBase.ConfigTools.find1_instance

PsychometricsBazaarBase.ConfigTools.find1_type

PsychometricsBazaarBase.ConfigTools.find1_type_sloppy

PsychometricsBazaarBase.ConfigTools.@requiresome

PsychometricsBazaarBase.ConfigTools.@returnsome

PsychometricsBazaarBase.ConfigTools.@returnsome