IntegralCoeffs

PsychometricsBazaarBase.IntegralCoeffs — Module

These are helpers for weighting integrals of p.d.f.s for calculating basic stats.

The main idea of doing it this way is to have a single instance of these to reuse specializations and to use structs so as to be able to control the level of specialization.

PsychometricsBazaarBase.IntegralCoeffs.OuterProdDev — Type

PsychometricsBazaarBase.IntegralCoeffs.PriorApply — Type

PsychometricsBazaarBase.IntegralCoeffs.SqDev — Type

PsychometricsBazaarBase.IntegralCoeffs.id — Method

id(x)

PsychometricsBazaarBase.IntegralCoeffs.one — Method

one(x_)

Index

PsychometricsBazaarBase.IntegralCoeffs

PsychometricsBazaarBase.IntegralCoeffs.OuterProdDev

PsychometricsBazaarBase.IntegralCoeffs.PriorApply

PsychometricsBazaarBase.IntegralCoeffs.SqDev

PsychometricsBazaarBase.IntegralCoeffs.id

PsychometricsBazaarBase.IntegralCoeffs.one