Integrators

PsychometricsBazzarBase.Integrators — Module

This module provides a common interface to different numerical integration techniques.

PsychometricsBazzarBase.Integrators.BareIntegrationResult — Type

The result of an integration technique which provides no error value.

vec::Any

PsychometricsBazzarBase.Integrators.CubaCuhre — Type

The Cuhre algorithm.

struct CubaCuhre <: PsychometricsBazzarBase.Integrators.CubaAlgorithm

PsychometricsBazzarBase.Integrators.CubaDivonne — Type

The Divonne algorithm.

struct CubaDivonne <: PsychometricsBazzarBase.Integrators.CubaAlgorithm

PsychometricsBazzarBase.Integrators.CubaIntegrator — Type

CubaIntegrator is a wrapper around the Cuba.jl integration functions.

struct CubaIntegrator{AlgorithmT<:PsychometricsBazzarBase.Integrators.CubaAlgorithm, KwargsT} <: PsychometricsBazzarBase.Integrators.Integrator

lo::Vector{Float64}

hi::Vector{Float64}

algorithm::PsychometricsBazzarBase.Integrators.CubaAlgorithm

kwargs::Any

Usage example:

CubaIntegrator([0.0, 0.0], [1.0, 1.0], CubaVegas()) do x
    x[1] * x[2]
end

PsychometricsBazzarBase.Integrators.CubaIntegrator — Method

(integrator::CubaIntegrator)(f[, ncomp, lo, hi; kwargs...])

Perform a Cuba integration.

PsychometricsBazzarBase.Integrators.CubaSuave — Type

The Sauve algorithm.

struct CubaSuave <: PsychometricsBazzarBase.Integrators.CubaAlgorithm

PsychometricsBazzarBase.Integrators.CubaVegas — Type

The VEGAS algorithm.

struct CubaVegas <: PsychometricsBazzarBase.Integrators.CubaAlgorithm

PsychometricsBazzarBase.Integrators.CubatureIntegrator — Type

Construct a Cubature integrator based on Cubature.jl with on a specified interval.

lo::Vector{Float64}

hi::Vector{Float64}

kwargs::Any

PsychometricsBazzarBase.Integrators.CubatureIntegrator — Method

(integrator::CubatureIntegrator)(f[, ncomp, lo, hi; kwargs...])

Perform a Cubature integration based on Cubature.jl.

PsychometricsBazzarBase.Integrators.ErrorIntegrationResult — Type

The result of an integration technique which provides an error value. Note that error values are not comparible between different integration techniques in general.

vec::Any

err::Any

PsychometricsBazzarBase.Integrators.FixedGKIntegrator — Type

Construct a fixed-order Gauss-Kronrod integrator based on QuadGK.jl with on a specified interval.

lo::Float64

hi::Float64

order::Int64

PsychometricsBazzarBase.Integrators.FixedGKIntegrator — Method

(integrator::QuadGKIntegrator)(f[, ncomp, lo, hi; order=...])

Perform a fixed-order Gauss-Kronrod integration based on QuadGK.jl.

PsychometricsBazzarBase.Integrators.HCubatureIntegrator — Type

Construct a Cubature integrator based on HCubature.jl with on a specified interval.

lo::Vector{Float64}

hi::Vector{Float64}

kwargs::Any

PsychometricsBazzarBase.Integrators.HCubatureIntegrator — Method

(integrator::HCubatureIntegrator)(f[, ncomp, lo, hi; kwargs...])

Perform Cubature integration based on HCubature.jl.

PsychometricsBazzarBase.Integrators.MCIVegasIntegrator — Type

Construct a VEGAS integrator based on MonteCarloIntegration.jl with on a specified interval.

lo::Vector{Float64}

hi::Vector{Float64}

kwargs::NamedTuple

PsychometricsBazzarBase.Integrators.MCIVegasIntegrator — Method

(integrator::HCubatureIntegrator)(f[, ncomp, lo, hi; kwargs...])

Perform a VEGAS integration based on MonteCarloIntegration.jl.

PsychometricsBazzarBase.Integrators.MultiDimFixedGKIntegrator — Type

Construct a fixed-order multi-dimensional Gauss-Kronrod integrator based on QuadGK.jl with on a specified interval.

lo::Vector{Float64}

hi::Vector{Float64}

order::AbstractVector{Int64}

PsychometricsBazzarBase.Integrators.MultiDimFixedGKIntegrator — Method

(integrator::QuadGKIntegrator)(f[, ncomp, lo, hi; order=...])

Perform a fixed-order multi-dimensional Gauss-Kronrod integrator based on QuadGK.jl.

PsychometricsBazzarBase.Integrators.QuadGKIntegrator — Type

Construct a adaptive Gauss-Kronrod integrator based on QuadGK.jl with on a specified interval.

lo::Float64

hi::Float64

order::Int64

PsychometricsBazzarBase.Integrators.QuadGKIntegrator — Method

(integrator::QuadGKIntegrator)(f[, ncomp, lo, hi; order=..., rtol=...])

Perform an adaptive Gauss-Kronrod integration using QuadGK.jl.

PsychometricsBazzarBase.Integrators.interr — Function

Given any integration result, get the integral error. In case the integration technique does not supply one, this returns nothing.

PsychometricsBazzarBase.Integrators.intval — Method

Given any integration result, get the integral value.

intval(res)

Index

PsychometricsBazzarBase.Integrators

PsychometricsBazzarBase.Integrators.BareIntegrationResult

PsychometricsBazzarBase.Integrators.CubaCuhre

PsychometricsBazzarBase.Integrators.CubaDivonne

PsychometricsBazzarBase.Integrators.CubaIntegrator

PsychometricsBazzarBase.Integrators.CubaIntegrator

PsychometricsBazzarBase.Integrators.CubaSuave

PsychometricsBazzarBase.Integrators.CubaVegas

PsychometricsBazzarBase.Integrators.CubatureIntegrator

PsychometricsBazzarBase.Integrators.CubatureIntegrator

PsychometricsBazzarBase.Integrators.ErrorIntegrationResult

PsychometricsBazzarBase.Integrators.FixedGKIntegrator

PsychometricsBazzarBase.Integrators.FixedGKIntegrator

PsychometricsBazzarBase.Integrators.HCubatureIntegrator

PsychometricsBazzarBase.Integrators.HCubatureIntegrator

PsychometricsBazzarBase.Integrators.MCIVegasIntegrator

PsychometricsBazzarBase.Integrators.MCIVegasIntegrator

PsychometricsBazzarBase.Integrators.MultiDimFixedGKIntegrator

PsychometricsBazzarBase.Integrators.MultiDimFixedGKIntegrator

PsychometricsBazzarBase.Integrators.QuadGKIntegrator

PsychometricsBazzarBase.Integrators.QuadGKIntegrator

PsychometricsBazzarBase.Integrators.interr

PsychometricsBazzarBase.Integrators.intval