Integrators

PsychometricsBazzarBase.ConfigTools

PsychometricsBazzarBase.IntegralCoeffs

PsychometricsBazzarBase.Integrators

PsychometricsBazzarBase.Optimizers

PsychometricsBazzarBase.Integrators.BareIntegrationResult

PsychometricsBazzarBase.Integrators.CubaCuhre

PsychometricsBazzarBase.Integrators.CubaDivonne

PsychometricsBazzarBase.Integrators.CubaIntegrator

PsychometricsBazzarBase.Integrators.CubaIntegrator

PsychometricsBazzarBase.Integrators.CubaSuave

PsychometricsBazzarBase.Integrators.CubaVegas

PsychometricsBazzarBase.Integrators.CubatureIntegrator

PsychometricsBazzarBase.Integrators.CubatureIntegrator

PsychometricsBazzarBase.Integrators.ErrorIntegrationResult

PsychometricsBazzarBase.Integrators.FixedGKIntegrator

PsychometricsBazzarBase.Integrators.FixedGKIntegrator

PsychometricsBazzarBase.Integrators.HCubatureIntegrator

PsychometricsBazzarBase.Integrators.HCubatureIntegrator

PsychometricsBazzarBase.Integrators.MCIVegasIntegrator

PsychometricsBazzarBase.Integrators.MCIVegasIntegrator

PsychometricsBazzarBase.Integrators.MultiDimFixedGKIntegrator

PsychometricsBazzarBase.Integrators.MultiDimFixedGKIntegrator

PsychometricsBazzarBase.Integrators.QuadGKIntegrator

PsychometricsBazzarBase.Integrators.QuadGKIntegrator

PsychometricsBazzarBase.Integrators.interr

PsychometricsBazzarBase.Integrators.intval

PsychometricsBazzarBase.ConfigTools.@requiresome

PsychometricsBazzarBase.ConfigTools.@returnsome

PsychometricsBazzarBase.Integrators — Module

This module provides a common interface to different numerical integration techniques.

PsychometricsBazzarBase.Integrators.BareIntegrationResult — Type

The result of an integration technique which provides no error value.

PsychometricsBazzarBase.Integrators.CubaCuhre — Type

The Cuhre algorithm.

PsychometricsBazzarBase.Integrators.CubaDivonne — Type

The Divonne algorithm.

PsychometricsBazzarBase.Integrators.CubaIntegrator — Type

CubaIntegrator is a wrapper around the Cuba.jl integration functions.

Usage example:

julia CubaIntegrator([0.0, 0.0], [1.0, 1.0], CubaVegas()) do x x[1] * x[2] end`

PsychometricsBazzarBase.Integrators.CubaIntegrator — Method

Perform a Cuba integration.

PsychometricsBazzarBase.Integrators.CubaSuave — Type

The Sauve algorithm.

PsychometricsBazzarBase.Integrators.CubaVegas — Type

The VEGAS algorithm.

PsychometricsBazzarBase.Integrators.CubatureIntegrator — Type

Construct a Cubature integrator based on Cubature.jl with on a specified interval.

PsychometricsBazzarBase.Integrators.CubatureIntegrator — Method

Perform a Cubature integration based on Cubature.jl.

PsychometricsBazzarBase.Integrators.ErrorIntegrationResult — Type

The result of an integration technique which provides an error value. Note that error values are not comparible between different integration techniques in general.

PsychometricsBazzarBase.Integrators.FixedGKIntegrator — Type

Construct a fixed-order Gauss-Kronrod integrator based on QuadGK.jl with on a specified interval.

PsychometricsBazzarBase.Integrators.FixedGKIntegrator — Method

Perform a fixed-order Gauss-Kronrod integration based on QuadGK.jl.

PsychometricsBazzarBase.Integrators.HCubatureIntegrator — Type

Construct a Cubature integrator based on HCubature.jl with on a specified interval.

PsychometricsBazzarBase.Integrators.HCubatureIntegrator — Method

Perform Cubature integration based on HCubature.jl.

PsychometricsBazzarBase.Integrators.MCIVegasIntegrator — Type

Construct a VEGAS integrator based on MonteCarloIntegration.jl with on a specified interval.

PsychometricsBazzarBase.Integrators.MCIVegasIntegrator — Method

Perform a VEGAS integration based on MonteCarloIntegration.jl.

PsychometricsBazzarBase.Integrators.MultiDimFixedGKIntegrator — Type

Construct a fixed-order multi-dimensional Gauss-Kronrod integrator based on QuadGK.jl with on a specified interval.

PsychometricsBazzarBase.Integrators.MultiDimFixedGKIntegrator — Method

Perform a fixed-order multi-dimensional Gauss-Kronrod integrator based on QuadGK.jl.

PsychometricsBazzarBase.Integrators.QuadGKIntegrator — Type

Construct a adaptive Gauss-Kronrod integrator based on QuadGK.jl with on a specified interval.

PsychometricsBazzarBase.Integrators.QuadGKIntegrator — Method

Perform an adaptive Gauss-Kronrod integration using QuadGK.jl.

PsychometricsBazzarBase.Integrators.interr — Function

Given any integration result, get the integral error. In case the integration technique does not supply one, this returns nothing.

PsychometricsBazzarBase.Integrators.intval — Method

Given any integration result, get the integral value.